Switch passing weather data to views to an associated value on HomeState.success.
Remove the Published weather property from the view model and also the placeholder value since neither are necessary. It means only the weather data is passed to WeatherView rather than the whole view model. Add a preview for HomeWeather since it’s needed in the view instead of the view model.