REFACTOR: Remove location access from view layer
The app was responding to system location authorization and fetching
location data from an ObservableObject connected to a root view before
passing that data back through all layers of the app to fetch weather
data remotely.

This update creates a new location service that handles all interactions
with CoreLocation in the framework layer of the app. Location data is no
longer passed through the app, instead is retrieved and used in the
appropriate places to support a Clean Architecture.